Move classes from Util to Crypt
