+++ www/core/Crypt/OpaqueIdentifier.php
@@ -19,9 +19,8 @@
  * Software Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-namespace SimpleID\Util;
+namespace SimpleID\Crypt;
 
-use SimpleID\Crypt\Random;
 use SimpleID\Store\StoreManager;
 
 /**
